% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_file_metadata.R
\name{analyze_file_metadata}
\alias{analyze_file_metadata}
\title{Analyze the metadata of a file}
\usage{
analyze_file_metadata(read_df, metadata_dir = NULL, save_file = TRUE,
  see_log = TRUE, create_log = TRUE, log_prefix = NA, log_id = NA,
  my_treshold = futile.logger::INFO)
}
\arguments{
\item{read_df}{A data frame read in by \code{\link{zacat_id_data}}}

\item{see_log}{\code{TRUE} which will print messages to the screen.}

\item{create_log}{It will create log files in the sr_logs director.}

\item{log_prefix}{Defaults to \code{NA}, in which case a
new one will be assigned to the logs (if they are requested.) The
log_prefix can be directly assigned.}

\item{log_id}{Defaults to \code{NA}, in which case a
new one will be assigned to the logs (if they are requested.) The
log_prefix can be directly assigned.}

\item{my_treshold}{Can be \code{futile.logger::WARN},
\code{futile.logger::INFO}, \code{futile.logger::ERROR}.Defaults to
\code{futile.logger::INFO}.}
}
\description{
Tries to analye the metadata of data frame read by \code{\link{zacat_id_data}}.
}
\examples{
\dontrun{
##use your own file path:
##The log parameters can be passed on optionally as ...
analyze_file_metadata(
     metadata = NULL,
     read_df = read_df,
     see_log = see_log,
     create_log = creat_log,
     my_treshold = my_treshold )
}
}
