% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_gesis_file.R
\name{analyze_gesis_file}
\alias{analyze_gesis_file}
\title{Analyze a GESIS file}
\usage{
analyze_gesis_file(gesis_file, see_log = TRUE, create_log = TRUE,
  log_prefix = NA, log_id = NA, my_treshold = futile.logger::INFO)
}
\arguments{
\item{gesis_file}{The full path to the GESIS SPSS file or a data.frame
read by haven consisting the contents of the file.}

\item{see_log}{\code{TRUE} which will print messages to the screen.}

\item{create_log}{It will create log files in the sr_logs director.}

\item{log_prefix}{Defaults to \code{NA}, in which case a
new one will be assigned to the logs (if they are requested.) The
log_prefix can be directly assigned.}

\item{log_id}{Defaults to \code{NA}, in which case a
new one will be assigned to the logs (if they are requested.) The
log_prefix can be directly assigned.}

\item{my_treshold}{Can be \code{futile.logger::WARN},
\code{futile.logger::INFO}, \code{futile.logger::ERROR}.Defaults to
\code{futile.logger::INFO}.}
}
\description{
There is a rather cumbersome exception handling with not-unique
GESIS variable names, which occur when certain questions are only asked
in some geographical units, such as the Turkish community of Cyprus.
In these cases, GESIS may or may not give unique variable names. Some
futher problems may arrise, currently the code detects this problem.
}
\details{
Runing the analyitcs function is very resource intensive. For this reason
the end result is always saved to the \code{tempdir()} which is deleted
automatically at the end of each session.  You can retrieve the metadata
from here, should you need it during an interactive session.
}
\examples{
\dontrun{
##use your own file path:
analyse_gesis_file( path = " ... your own file path ... ",
                   see_log = TRUE,
                   create_log = TRUE,
                   my_treshold = futile.logger::INFO)
}
}
