% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_character.factor_pos_neg.R
\name{as_character.factor_pos_neg}
\alias{as_character.factor_pos_neg}
\title{Convert to character from a factor_pos_neg object}
\usage{
\method{as_character}{factor_pos_neg}(x, type = NULL)
}
\arguments{
\item{x}{An object to be converted to character}

\item{type}{Default to \code{NULL}. Not used in this instance
of the method.}
}
\description{
Convert to character from a factor_pos_neg object
}
\examples{
my_categories <- as_factor_pos_neg(
   c("Better", "DK", "Worse",
     "Same", "The Same", "Inap. not")
     )

as_character (my_categories)

}
