% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_character.R
\name{as_character}
\alias{as_character}
\title{Convert to character}
\usage{
as_character(x, type = NULL)
}
\arguments{
\item{x}{An object to be converted}

\item{type}{If necessary, the type of conversion}
}
\description{
Convert to character
}
\examples{
my_binary <- as_factor_binary(c("mentioned", "mentioned",
 "not mentioned", "dk"))

as_character (my_binary)

}
