% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_factor.factor_3.R
\name{as_factor.factor_3}
\alias{as_factor.factor_3}
\title{Convert to factor from a factor_3 object}
\usage{
\method{as_factor}{factor_3}(x, type = NULL)
}
\arguments{
\item{x}{An object to be converted to factor}

\item{type}{Default to \code{NULL}. Not used in this instance
of the method.}
}
\description{
Convert to factor from a factor_3 object
}
\examples{
my_categories <- as_factor_3( c("Rural area or village",
 "Small or medium-sized town", "Inap.", "Large town/city",
  "DK", "Large town/city"))

as_factor (my_categories)

}
