% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_factor.factor_4.R
\name{as_factor.factor_4}
\alias{as_factor.factor_4}
\title{Convert to factor from a factor_4 object}
\usage{
\method{as_factor}{factor_4}(x, type = NULL)
}
\arguments{
\item{x}{An object to be converted to factor}

\item{type}{Default to \code{NULL}. Not used in this instance
of the method.}
}
\description{
Convert to factor from a factor_4 object
}
\examples{
my_categories <- as_factor_4( c("Not at all", "Low",
               "Medium", "Strong", "DK"))

as_factor (my_categories)

}
