% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_factor.factor_binary.R
\name{as_factor.factor_binary}
\alias{as_factor.factor_binary}
\title{Convert to factor from a factor_binary object}
\usage{
\method{as_factor}{factor_binary}(x, type = NULL)
}
\arguments{
\item{x}{An object to be converted}

\item{type}{Default to \code{NULL}. Not used in this instance
of the method.}
}
\description{
Convert to factor from a factor_binary object
}
\examples{
my_binary <- as_factor_binary(c("mentioned", "mentioned",
 "not mentioned", "dk"))

as_factor (my_binary)

}
