% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_factor.R
\name{as_factor}
\alias{as_factor}
\title{Convert to factor}
\usage{
as_factor(x, type = NULL)
}
\arguments{
\item{x}{An object to be converted}

\item{type}{If necessary, the type of conversion}
}
\description{
Convert to factor
}
\examples{
my_binary <- as_factor_binary(c("mentioned", "mentioned",
 "not mentioned", "dk"))

as_factor (my_binary)

}
