% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_factor_3.R
\name{as_factor_3}
\alias{as_factor_3}
\title{Constructor for factor_3 three positive-level factor}
\usage{
as_factor_3(x)
}
\arguments{
\item{x}{An input vector to be converted to a
three positive level variable}
}
\description{
Constructor for factor_3 three positive-level factor
}
\examples{
as_factor_3( c("Rural area or village", "Small or medium-sized town",
                  "Inap.", "Large town/city", "DK", "Large town/city"))

}
