% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_factor_4.R
\name{as_factor_4}
\alias{as_factor_4}
\title{Constructor for factor_4 four non-negative level factor}
\usage{
as_factor_4(x)
}
\arguments{
\item{x}{An input vector to be converted to a
four non-negative level variable}
}
\description{
Constructor for factor_4 four non-negative level factor
}
\examples{
as_factor_4( c("Not at all", "Low",
               "Medium", "Strong", "DK"))

}
