% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_factor_5.R
\name{as_factor_5}
\alias{as_factor_5}
\title{Constructor for factor_5 four non-negative level factor}
\usage{
as_factor_5(x)
}
\arguments{
\item{x}{An input vector to be converted to a
four non-negative level variable}
}
\description{
Constructor for factor_5 four non-negative level factor
}
\examples{
as_factor_5( c("Much lower", "Somewhat lower",
	"More or less the same", 	"Somewhat higher", "Much higher"))

}
