% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_factor_binary.R
\name{as_factor_binary}
\alias{as_factor_binary}
\title{Constructor for factor_binary}
\usage{
as_factor_binary(x)
}
\arguments{
\item{x}{An input vector to be converted to binary factor class.}
}
\description{
Constructor for factor_binary
}
\examples{
as_factor_binary(c("mentioned", "mentioned",
                   "not mentioned", "dk"))

as_factor_binary ( c("yes", "no", "DK", "Inap."))

}
