% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_factor_frequency.R
\name{as_factor_frequency}
\alias{as_factor_frequency}
\title{Constructor for quasi-numeric frequency factors}
\usage{
as_factor_frequency(x)
}
\arguments{
\item{x}{An input vector to be converted to time frequencies}
}
\description{
Constructor for quasi-numeric frequency factors
}
\examples{
as_factor_frequency( c("Less often", "Several times a month",
               "Every 2-3 months", "Once a month", "DK"))

}
