% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_factor_pos_neg.R
\name{as_factor_pos_neg}
\alias{as_factor_pos_neg}
\title{Constructor for factor_yes_no_4}
\usage{
as_factor_pos_neg(x)
}
\arguments{
\item{x}{An input vector to be converted to a four level yes-no variable}
}
\description{
Constructor for factor_yes_no_4
}
\examples{
as_factor_pos_neg( c("Better", "DK", "Worse",
                   "Same", "The Same", "Inap. not"))

}
