% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_factor_yes_no_4.R
\name{as_factor_yes_no_4}
\alias{as_factor_yes_no_4}
\title{Constructor for factor_yes_no_4}
\usage{
as_factor_yes_no_4(x)
}
\arguments{
\item{x}{An input vector to be converted to a four level yes-no variable}
}
\description{
Constructor for factor_yes_no_4
}
\examples{
as_factor_yes_no_4( c("Very good", "Rather Good",
                  "Inap.", "Rather Bad", "DK", "Very bad"))

}
