% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_numeric.factor_4.R
\name{as_numeric.factor_4}
\alias{as_numeric.factor_4}
\title{Convert to numeric from a factor_4 object}
\usage{
\method{as_numeric}{factor_4}(x, type = NULL)
}
\arguments{
\item{x}{An object to be converted to numeric}

\item{type}{Default to \code{NULL}. Not used in this instance
of the method.}
}
\description{
Convert to numeric from a factor_4 object
}
\examples{
my_categories <- as_factor_4( c("Not at all", "Low",
               "Medium", "Strong", "DK"))

as_numeric (my_categories)

}
