% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_numeric.factor_pos_neg.R
\name{as_numeric.factor_pos_neg}
\alias{as_numeric.factor_pos_neg}
\title{Convert to numeric from a factor_pos_neg object}
\usage{
\method{as_numeric}{factor_pos_neg}(x, type = NULL)
}
\arguments{
\item{x}{An object to be converted to numeric}

\item{type}{Default to \code{NULL}. Not used in this instance
of the method.}
}
\description{
Convert to numeric from a factor_pos_neg object
}
\examples{
my_categories <- as_factor_pos_neg( c("Very good", "Rather Good",
                  "Inap.", "Rather Bad", "DK", "Very bad"))

as_numeric (my_categories)

}
