% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_numeric.labelled.R
\name{as_numeric.labelled}
\alias{as_numeric.labelled}
\title{Convert to numeric from a GESIS labelled SPSS object}
\usage{
\method{as_numeric}{labelled}(x, type = "as_factor_binary")
}
\arguments{
\item{x}{An object to be converted to numeric}

\item{type}{Any of \code{as_factor_binary}, \code{as_factor_3},
 \code{as_factor_pos_neg}, \code{as_factor_yes_no_4},
\code{rescale_time_frequency_use}, \code{rescale_alphanumeric}.}
}
\description{
Convert to numeric from a GESIS labelled SPSS object
}
\examples{
my_categories <- as_factor_3( c("Rural area or village",
 "Small or medium-sized town", "Inap.", "Large town/city",
  "DK", "Large town/city"))

as_numeric (my_categories)

}
