% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_numeric.R
\name{as_numeric}
\alias{as_numeric}
\title{Convert to numeric}
\usage{
as_numeric(x, type = NULL)
}
\arguments{
\item{x}{An object to be converted}

\item{type}{If necessary, the type of conversion}
}
\description{
Convert to numeric
}
\examples{
my_binary <- as_factor_binary(c("mentioned", "mentioned",
 "not mentioned", "dk"))

as_numeric (my_binary)

}
