% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_values_get.R
\name{binary_values_get}
\alias{binary_values_get}
\title{Get binary yes values}
\usage{
binary_values_get()
}
\description{
Get the character strings that will be recoded as "yes" in
binary variables.
}
\examples{
binary_values_get()

}
