% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr_logical.R
\name{chr_logical}
\alias{chr_logical}
\title{Character to logical conversion}
\usage{
chr_logical(column, false_id = "default", true_id = "default",
  na_labels = "default", na_value = NA)
}
\arguments{
\item{column}{A column from a survey data frame where gender is recorded.}

\item{false_id}{The false values in the original raw data. defaults to \code{"0"}.
\code{"default"} searches for statements starting with \code{c("nincs", "none", "ne", "no", "non", "never")}.}

\item{true_id}{The true values in the original raw data, defaults to \code{"1"}.
\code{"default"} searches for statements starting with \code{c("van", "igen", "yes", "da", "oui")}.}

\item{na_labels}{Is there a special character ID for missing variables? (Default: \code{"default"} )}

\item{na_value}{What should be the value of missing answers (Default: \code{"NA"}, but in some cases
it is useful to code them 0. )
If you get a warning with NA conversion, you are likely to have several NA characters,
such as \code{missing} and \code{NA} at the same time.}
}
\description{
Converts logical answers from text documents.
}
\examples{
chr_logical(column = as.character (c("1 - yes", "1", "0 - no",
           "missing", NA)),
           false_id  = "0 - no",
           true_id = "1 - yes",
           na_labels = "missing" )
}
