% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr_numeric.R
\name{chr_numeric}
\alias{chr_numeric}
\title{Charcter to numeric conversion}
\usage{
chr_numeric(column, validate_min = NA, validate_max = NA, digits = NA,
  decimals = "point", na_id = "NA", is_integer = FALSE)
}
\arguments{
\item{column}{A column from a survey data frame where gender is recorded.}

\item{validate_min}{The valid minimum value. Defaults to \code{NA} in which case no validation takes place.}

\item{validate_max}{The valid maximum value.  Defaults to \code{NA} in which case no validation takes place.}

\item{digits}{How many digits are valid in the raw date. Defaults to \code{NA} when
digits are not truncated. \code{2} uses only the first two characters of the raw data.
Digits are counted as characters. So keeping 1.5 requires \code{digit = 3}.}

\item{decimals}{Defaults to \code{"point"}, alternative is \code{"comma"}}

\item{na_id}{Is there a special character ID for missing variables? (Default: \code{"NA"} )}

\item{is_integer}{Numeric or integer should be returned?  Defaults to \code{FALSE} in which case no validation takes place.
If you get a warning with NA conversion, you are likely to have several NA characters,
such as \code{missing} and \code{NA} at the same time.}
}
\description{
Convert a character to a numeric expression.
}
\examples{
chr_numeric(column = as.character (c(1982, 1990.5, 2019,
                                     "missing", NA, 1678)),
            validate_min  = 1900, validate_max = 2017,
            digits = NA, decimals = "point",
            na_id = "missing",
            is_integer = FALSE )
chr_numeric(column = as.character (c("1 - egyet", "11 - tizenegyet",
            "missing", "2")),
            validate_min  = NA, validate_max = NA,
            digits = 2,
            na_id = "missing",
            is_integer = FALSE )
}
