% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_conversion_suggest.R
\name{class_conversion_suggest}
\alias{class_conversion_suggest}
\title{Suggest variable conversion}
\usage{
class_conversion_suggest(x)
}
\arguments{
\item{x}{A variable from a GESIS archive.}
}
\description{
Suggest variable conversion
}
\examples{
\dontrun{
suggest_variable_conversion (x)
}
}
