% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_numeric.R
\name{convert_to_numeric}
\alias{convert_to_numeric}
\title{Type conversion on a labelled data frame}
\usage{
convert_to_numeric(df, metadata, conversion_type = "factor_binary",
  see_log = TRUE, create_log = TRUE, log_prefix = NA, log_id = NA,
  my_treshold = futile.logger::INFO)
}
\arguments{
\item{df}{A labelled data.frame to be converted to numeric}

\item{metadata}{the metadata}

\item{conversion_type}{Defaults to  \code{"factor_binary"}}

\item{see_log}{\code{TRUE} which will print messages to the screen.}

\item{create_log}{It will create log files in the sr_logs director.}

\item{log_prefix}{Defaults to \code{NA}, in which case a
new one will be assigned to the logs (if they are requested.) The
log_prefix can be directly assigned.}

\item{log_id}{Defaults to \code{NA}, in which case a
new one will be assigned to the logs (if they are requested.) The
log_prefix can be directly assigned.}

\item{my_treshold}{Can be \code{futile.logger::WARN},
\code{futile.logger::INFO}, \code{futile.logger::ERROR}.Defaults to
\code{futile.logger::INFO}.}
}
\description{
Type conversion on a labelled data frame
}
\examples{
\dontrun{
conversion_function( df = read_df2, metadata, see_log = see_log,
create_log = create_log,
log_prefix = log_prefix,
log_id = log_id,
my_treshold = my_treshold)
}
}
