% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurobarometer.R
\docType{package}
\name{eurobarometer}
\alias{eurobarometer}
\alias{eurobarometer-package}
\title{\code{eurobarometer} package}
\description{
Eurobarometer survey harmonization package
}
\details{
See the README on
\href{https://github.com/antaldaniel/eurobarometer#readme}{GitHub}
}
