% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_3_get.R
\name{factor_3_get}
\alias{factor_3_get}
\title{Get three positive level categories}
\usage{
factor_3_get()
}
\description{
Get the character strings that will be represented by \code{0},
\code{1} or \code{2}.
}
\examples{
factor_3_get()

}
