% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_binary_get.R
\name{factor_binary_get}
\alias{factor_binary_get}
\title{Get four-level yes values}
\usage{
factor_binary_get()
}
\description{
Get the character strings that will be recoded as "yes" in
four-level variables.
}
\examples{
yes_4_get

}
