% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_pos_neg_get.R
\name{factor_pos_neg_get}
\alias{factor_pos_neg_get}
\title{Get three level positive, zero, negative factors}
\usage{
factor_pos_neg_get()
}
\description{
Get the character strings that will be represented by \code{-1},
\code{0} or \code{1}.
}
\examples{
factor_pos_neg_get()

}
