% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesis_file_read.R
\name{gesis_file_read}
\alias{gesis_file_read}
\title{Read an convert a GESIS file}
\usage{
gesis_file_read(zacat_id = "ZA5688", data_dir = "data-raw/",
  metadata_dir = NULL, new_names = NULL, conversion = "numeric",
  save_file = TRUE, see_log = TRUE, create_log = TRUE, log_prefix = NA,
  log_id = NA, my_treshold = futile.logger::INFO)
}
\arguments{
\item{zacat_id}{\code{"ZA5688"}}

\item{data_dir}{\code{"data-raw/"}}

\item{metadata_dir}{\code{NULL}}

\item{new_names}{\code{NULL}}

\item{conversion}{\code{"numeric"}}

\item{save_file}{\code{TRUE}}

\item{see_log}{\code{TRUE} which will print messages to the screen.}

\item{create_log}{It will create log files in the sr_logs director.}

\item{log_prefix}{Defaults to \code{NA}, in which case a
new one will be assigned to the logs (if they are requested.) The
log_prefix can be directly assigned.}

\item{log_id}{Defaults to \code{NA}, in which case a
new one will be assigned to the logs (if they are requested.) The
log_prefix can be directly assigned.}

\item{my_treshold}{Can be \code{futile.logger::WARN},
\code{futile.logger::INFO}, \code{futile.logger::ERROR}.Defaults to
\code{futile.logger::INFO}.}
}
\description{
Runing the analyitcs function is very resource intensive. For this reason
the end result is always saved to the \code{tempdir()} which is deleted
automatically at the end of each session.  You can retrieve the metadata
from here, should you need it during an interactive session.
}
\examples{
\dontrun{
##use your own file path:
analyse_gesis_file( path = " ... your own file path ... ",
                   see_log = TRUE,
                   create_log = TRUE,
                   my_treshold = futile.logger::INFO)
}
}
