% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesis_longname_harmonization.R
\name{gesis_longname_harmonization}
\alias{gesis_longname_harmonization}
\title{GESIS long name harmonization}
\usage{
gesis_longname_harmonization(long_name)
}
\arguments{
\item{long_name}{The updated metadata$suggested_name}
}
\description{
Using the data in the short variable names (codes) to correctly assign a
long variable name
}
\examples{
\dontrun{
 gesis_shortname_harmonization (long_name  = "Weight extrapolated 15+")
}
}
