% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesis_metadata_get.R
\name{gesis_metadata_get}
\alias{gesis_metadata_get}
\title{Get a metadata file}
\usage{
gesis_metadata_get(zacat_id = NULL, data_dir = NULL)
}
\arguments{
\item{zacat_id}{For example, \code{"ZA5688_v6-0-0"}, as analyzed and saved by
\code{\link{analyze_gesis_file}}.}

\item{data_dir}{Defaults to \code{"NULL"}. In this case the \code{"tempdir()"}
of your current R session will be used.}
}
\description{
Retrieve a metadata file with from your data directory or the temporary directory
of your current R session.
}
\examples{
\dontrun{
##use your own file:
gesis_metadata_get( zacat_id = "ZA5688",
                    data_dir = "data-raw/")
}
}
