% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_answer_options.R
\name{get_answer_options}
\alias{get_answer_options}
\title{Get answer options from trend questions.}
\usage{
get_answer_options(language = "en")
}
\arguments{
\item{language}{Defaults to \code{"en"}. Limited ability to test
in \code{"hu"}.}
}
\description{
Get answer options from trend questions.
}
\examples{
get_answer_options()
}
