% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eb_questionnaire_item.R
\name{get_eb_questionnaire_item}
\alias{get_eb_questionnaire_item}
\title{A helper function to separate GESIS variable name parts.}
\usage{
get_eb_questionnaire_item(x)
}
\arguments{
\item{x}{A variable to summarize.}
}
\description{
A helper function to separate GESIS variable name parts.
}
\examples{
\dontrun{
get_eb_questionnaire_item <- function ( x )
}
}
