% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_na_labels.R
\name{get_na_labels}
\alias{get_na_labels}
\title{Get standard NA labels.}
\usage{
get_na_labels()
}
\description{
Get standard NA labels.
}
\examples{
get_na_labels()

##Create a custom NA label filter
my_na_filter <- c(get_na_labels()$default,
                  "Not_among_these")
cat(my_na_filter)
}
