% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_naming_exceptions.R
\name{get_naming_exceptions}
\alias{get_naming_exceptions}
\title{Get naming exceptions}
\usage{
get_naming_exceptions()
}
\description{
These names are not changed automatically with regex rules.
}
\examples{
get_naming_exceptions
}
