% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_nuts2_mean.R
\name{impute_nuts2_mean}
\alias{impute_nuts2_mean}
\title{Impute NUTS2 regional mean}
\usage{
impute_nuts2_mean(df, impute_countries = c("DE", "IT", "UK"),
  weight_var = NULL, nuts_code = "code2010")
}
\arguments{
\item{df}{column from a survey data frame where gender is recorded.}

\item{impute_countries}{The value labels}

\item{weight_var}{The weighting variable in \code{df}. Defaults to \code{NULL}.}

\item{nuts_code}{Currently only \code{code2010} and \code{code2013}
is supported.}
}
\description{
Impute NUTS2 regional mean
}
\examples{
sample  <- gesis_sample[,-c(1,3,10)]
impute_nuts2_mean(sample, nuts_code = "code2010")

}
