% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_nuts2_median.R
\name{impute_nuts2_median}
\alias{impute_nuts2_median}
\title{Impute NUTS2 regional mean}
\usage{
impute_nuts2_median(df, impute_countries = c("DE", "IT", "GB"),
  weight_var = NULL, nuts_code = "code2010")
}
\arguments{
\item{df}{column from a survey data frame where gender is recorded.}

\item{impute_countries}{The value labels}

\item{weight_var}{The weighting variable in \code{df}. Defaults to \code{NULL}.}

\item{nuts_code}{Currently only \code{code2010} and \code{code2013}
is supported.}
}
\description{
Impute NUTS2 regional mean
}
\examples{
sample  <- gesis_sample[, c(2 ,13, 11)]
sample$age_exact <- keep_numeric(sample$age_exact)
impute_nuts2_median(sample, nuts_code = "code2010")
}
