% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.factor_4.R
\name{is.factor_4}
\alias{is.factor_4}
\title{Is the object of class factor_4?}
\usage{
is.factor_4(x)
}
\arguments{
\item{x}{An object to be checked if it is a 4-level
non-negative categorical variable}
}
\description{
Is the object of class factor_4?
}
\examples{
my_categories <- as_factor_4( c("Not at all", "Low",
               "Medium", "Strong", "DK"))

is.factor_4  (my_categories)

}
