% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.factor_5.R
\name{is.factor_5}
\alias{is.factor_5}
\title{Is the object of class factor_5?}
\usage{
is.factor_5(x)
}
\arguments{
\item{x}{An object to be checked if it is a 5-level positive
categorical variable.}
}
\description{
Is the object of class factor_5?
}
\examples{
my_categories <- as_factor_5( c("Much lower", "Somewhat lower",
	"More or less the same", 	"Somewhat higher", "Much higher")

is.factor_5 (my_categories)

}
