% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.factor_binary.R
\name{is.factor_binary}
\alias{is.factor_binary}
\title{Is the object of class factor_binary?}
\usage{
is.factor_binary(x)
}
\arguments{
\item{x}{An object to be checked if it is a binary factor}
}
\description{
Is the object of class factor_binary?
}
\examples{
my_binary <- factor_binary(c("mentioned", "mentioned", "not mentioned", "dk"))

is.factor_binary ( my_binary)

}
