% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.factor_frequency.R
\name{is.factor_frequency}
\alias{is.factor_frequency}
\title{Is the object of class factor_frequency?}
\usage{
is.factor_frequency(x)
}
\arguments{
\item{x}{An object to be checked if it is a quasi-numeric frequency
categorical variable}
}
\description{
Is the object of class factor_frequency?
}
\examples{
my_categories <- as_factor_frequency(
 x  = c("Less often", "Several times a month",
 "Every 2-3 months", "Once a month", "DK"))

is.factor_frequency  (my_categories)

}
