% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.factor_pos_neg.R
\name{is.factor_pos_neg}
\alias{is.factor_pos_neg}
\title{Is the object of class factor_3?}
\usage{
is.factor_pos_neg(x)
}
\arguments{
\item{x}{An object to be checked if it is a 3-level positive
categorical variable.}
}
\description{
Is the object of class factor_3?
}
\examples{
my_categories <- as_factor_pos_neg( c("Rural area or village",
 "Small or medium-sized town", "Inap.", "Large town/city",
  "DK", "Large town/city"))

is.factor_pos_neg  (my_categories)

}
