% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.factor_yes_no_4.R
\name{is.factor_yes_no_4}
\alias{is.factor_yes_no_4}
\title{Is the object of class factor_yes_no_4?}
\usage{
is.factor_yes_no_4(x)
}
\arguments{
\item{x}{An object to be checked if it is a 4-level (2 positive,
2 negative) categorical variable.}
}
\description{
Is the object of class factor_yes_no_4?
}
\examples{
my_categories <- as_factor_yes_no_4( c("Very good", "Rather Good",
                  "Inap.", "Rather Bad", "DK", "Very bad"))

is.factor_yes_no_4 (my_binary)

}
