% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_question_viewed.R
\name{is_question_viewed}
\alias{is_question_viewed}
\title{Is question viewed?}
\usage{
is_question_viewed(column, na_value = 0)
}
\arguments{
\item{column}{A column from a survey data frame where gender is recorded.}

\item{na_value}{What should be the value of missing answers (Default: \code{"NA"}, but in some cases
it is useful to code them 0. )}
}
\description{
This is a (malformatted) logical variable used by the SurveyMonkey
Excel exporter. It is used to evaluate if a multiple choice question was
answered or not.
}
\examples{
is_question_viewed (column = c("Viewed", NA, "Any characters",
                        NA, NA),
                       na_value = 0 )
}
