% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_numeric.R
\name{keep_numeric}
\alias{keep_numeric}
\title{Keep numeric elements in a string}
\usage{
keep_numeric(column, na_labels = NULL, max_value = NULL, min_value = NULL,
  comma = FALSE)
}
\arguments{
\item{column}{column from a survey data frame where numbers are included with
units or other test, such as \code{"15 years old"}.}

\item{na_labels}{An explicit missing label to be removed such as "refusal".
The default value is \code{NULL}, but \code{"default"} will use a setting
\code{c("refusal", "nt/nv", "dk", "dk/na")} and remove all observations
starting with a label \code{Inap.}.}

\item{max_value}{Optional maximum value, anything reaching this level will be replaced with \code{NA}.}

\item{min_value}{Optional maximum value, anything reaching this level will be replaced with \code{NA}.}

\item{comma}{If decimals are written with commas, such as 10,7 instead of 10.7
Defaults to \code{FALSE}.}
}
\description{
Keep numeric elements in a string
}
\examples{
keep_numeric("He is 183.5 cm tall")

keep_numeric("He is 183,5 cm tall", comma=TRUE)

keep_numeric("He is 383.5 cm tall", max_value = 250)
}
