% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled_to_numeric.R
\name{labelled_to_numeric}
\alias{labelled_to_numeric}
\title{Labelled to numeric}
\usage{
labelled_to_numeric(labelled_var, labels, numeric_values)
}
\arguments{
\item{labelled_var}{column from a survey data frame where gender is recorded.}

\item{labels}{The value labels}

\item{numeric_values}{The numeric values}
}
\description{
Converts a standard better / same / worse scale to numeric 1, 0, -1.
}
