% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_item_identify.R
\name{questionnaire_item_identify}
\alias{questionnaire_item_identify}
\title{Identify questionnaire item}
\usage{
questionnaire_item_identify(x)
}
\arguments{
\item{x}{The original GESIS variable label (long)}
}
\description{
Identify questionnaire item
}
\examples{
\dontrun{
var_name_suggest ( gesis_name )
}
}
