% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_gender.R
\name{recode_gender}
\alias{recode_gender}
\title{Code gender variable}
\usage{
recode_gender(column, female_id = "default", female_value = 1,
  na_id = "NA", return_class = "numeric")
}
\arguments{
\item{column}{A column from a survey data frame where gender is recorded.}

\item{female_id}{A regex identifier for the female values. If omitted,
the default value is \code{"default"} searches for \code{"woman"} or
\code{"female"} (not case-sensitive).}

\item{female_value}{What should be the returned value for females (Default: \code{1} )}

\item{na_id}{Is there a special character ID for missing variables? (Default: \code{"NA"} )
Missing variables will be always treated as missing, however, if you have a special NA identifier,
It will also be treated as missing. Every other case will be treated "male"}

\item{return_class}{Default is \code{"numeric"}, alternatives \code{"character"} or
\code{"factor"}.}
}
\description{
Create a gender indicator or dummy from a character vector.
}
\examples{
code_gender(column = as.character (c("1-no", "1-nő",
                                     "NA", NA, "0")),
            female_id = "1", female_value = 1,
            na_id = "NA")
}
