% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_alphanumeric_en.R
\name{rescale_alphanumeric_en}
\alias{rescale_alphanumeric_en}
\title{Rescale alphanumeric variables in English}
\usage{
rescale_alphanumeric_en(column, from = NULL, to = NULL,
  na_labels = c("default"), return_class = "numeric")
}
\arguments{
\item{column}{A column from a survey data with a four-level agreement scale.}

\item{from}{Defaults to \code{c("NULL")} in which case the numbers 0-40
and 45 are coded, for example, from \code{"Twenty-one"}.}

\item{to}{Defaults to \code{c(0:40,41,NA)}.}

\item{na_labels}{Defaults to \code{c("default")}.}

\item{return_class}{Default is \code{"numeric"}, alternatives \code{"character"} or
\code{"factor"}.}
}
\description{
Create a numeric (or other) variable from alphanumeric coding, such as
"twenty", "twenty-one".\cr
This is a wrapper function around the\code{\link{rescale_categories}} function.
}
\examples{
rescale_alphanumeric_en  (column = c("One", "Twenty",
                           "twenty-one", "sEven", "NA"),
                          na_labels = "default")
}
