% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_categories.R
\name{rescale_categories}
\alias{rescale_categories}
\title{Rescale categories}
\usage{
rescale_categories(column, from = c("Agree", "Disagree"), to = c(1, 0),
  na_labels = NULL, exact_from = TRUE, return_class = "numeric")
}
\arguments{
\item{column}{A column from a survey data frame with categorical variables.}

\item{from}{Defaults to \code{c("Agree", "Disagree")}. Shorthened categories can be
used, to, in this case the first part of the character string is matched, for example,
\code{"Totally agree"} in case of \code{from = "totally"} is matched for the first
six characters,  \code{"not agree"} in case of \code{from = "not"}  is matched against
the first three characters.}

\item{to}{Defaults to \code{c(1,0)}.}

\item{na_labels}{Defaults to \code{NULL}. The pre-coded profiles can be accessed via
\code{\link{get_na_labels}}.}

\item{exact_from}{Deafults to \code{TRUE}. If \code{FALSE} you can use the
partial matching, but beware that in this case, \code{"twenty"} will be replaced by
\code{20}, and \code{"twenty-one"} will be already matched with \code{"twenty"}.}

\item{return_class}{Default is \code{"numeric"}, alternatives \code{"character"} or
\code{"factor"}.}
}
\description{
A generic function to rescale categorical variables. \cr
There are wrapper functions available for the Eurobarometer trend
and demography questions with the default GESIS codes:\cr
For example: \cr
\code{\link{rescale_marital_status}} for marital status. \cr
\code{\link{rescale_occupation}} for occupation and last job. \cr
The \code{\link{get_answer_options}} functions lists all the
currently recognized standard values.
}
\examples{
rescale_categories (column =
        c("Agree", "AGREE", "disagree",
          "DK", "Refusal (SPONT.)", "Disagree"),
        na_labels = c("DK", "Refusal (SPONT.)")
        )

rescale_categories (column =
        c("Agree", "AGREE", "disagree",
          "DK", "Refusal (SPONT.)", "Disagree"),
        na_labels = "default")

rescale_categories (column = c("Agree", "AGREE", "disagree"))

}
