% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_date_interview.R
\name{rescale_date_interview}
\alias{rescale_date_interview}
\title{Rescale date of interview}
\usage{
rescale_date_interview(column, return_class = "Date")
}
\arguments{
\item{column}{A column from a survey data where the date of interview is recorded.}

\item{return_class}{Default is \code{"Date"}, alternative is \code{"character"}.}
}
\description{
The function removes English day names, and English ordinal numbers such
as \code{1st}, \code{2nd}, \code{3rd}, \code{24th} are changed to cardinal
numbers \code{1}, \code{2}, \code{3}, \code{24}, etc.
}
\examples{
rescale_date_interview (column =
                          c("Saturday 26th April 2014",
                            "Monday 28th April 2014"),
                        return_class = "character")
}
