% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_marital_status.R
\name{rescale_marital_status}
\alias{rescale_marital_status}
\title{Rescale the Eurobarometer marital status variable.}
\usage{
rescale_marital_status(column, from = NULL, to = NULL,
  na_labels = c("default"), return_class = "factor", underscore = TRUE)
}
\arguments{
\item{column}{A column from a survey data with the subjective urbanization answers.}

\item{from}{Defaults to \code{c("NULL")} in which case the GESIS coding in English
\code{c("Widow/er: living without children","Single living w partner: without children" ...} will
be used.
The GESIS coding is not same across files.}

\item{to}{Defaults to \code{NULL} in which case the GESIS coding will be
shortened to \code{c("widow_no_children", "partners_no_children" ...)}.}

\item{na_labels}{Defaults to \code{c("default")}.}

\item{return_class}{Default is \code{"factor"} the alternative is \code{"character"}.}

\item{underscore}{Defaults to \code{TRUE} in which case factor names or character strings
contain underscore_between_words.  This is a better approach for further programming,
but you can choose \code{FALSE} for nicer printing results. See examples.}
}
\description{
If you only need a simpler categorization, use the 4- or 5- category
versions of this variable.
This is a truly categorical variable. The \code{return_class="factor"}
and \code{return_class="character"} give shorthened versions of the GESIS labels for easier handling in R.
This is a wrapper function around the \code{\link{rescale_categories}} function.
}
\examples{
rescale_marital_status (c("Widow/er: living without children",
                          "(Re-)Married: w children of this marriage",
                          "DK"),
                        return_class = "character",
                        underscore = FALSE)
}
