% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_occupation.R
\name{rescale_occupation}
\alias{rescale_occupation}
\title{Rescale the Eurobarometer occupation variable}
\usage{
rescale_occupation(column, from = NULL, to = NULL,
  na_labels = c("default"), return_class = "factor", underscore = TRUE)
}
\arguments{
\item{column}{A column from a survey data with the subjective urbanization answers.}

\item{from}{Defaults to \code{c("NULL")} in which case the GESIS coding in English
\code{c(..., "Retired, unable to work","Farmer", ...)} will
be used.}

\item{to}{Defaults to \code{NULL} in which case the GESIS coding will be
shortened to \code{c(... "retired", "farmer" ...)}. If you choose \code{return_class = "numeric"}
the default \code{to} parameter changes to \code{c(0,1,2)}.}

\item{na_labels}{Defaults to \code{c("default")}.}

\item{return_class}{Default is \code{"factor"}, alternatives \code{"character"} or
\code{"numeric"}.}

\item{underscore}{Defaults to \code{TRUE} in which case factor names or character strings
contain underscore_between_words.  This is a better approach for further programming,
but you can choose \code{FALSE} for nicer printing results. See examples.}
}
\description{
The Eurobarometer occupations have three groups: \code{none-active}, \code{self-employed},
and \code{employed} occupations. If you only need this categorization, you should
use the recatogorized version of this variable.
The numeric recategorization keeps this grouping, giving numeric representations
starting with digit \code{1} for \code{none-active}, \code{2} for \code{self-employed}
and \code{3} for \code{employed} respondents. Keeping only the first digit
will re-create the recoded version of this variable. The second digit is a unique
numeric identifier, but the ordering of the numeric variable, or the differences between
numeric codes have little if any meaning. This is a truly categorical variable.
The \code{return_class="factor"} and \code{return_class="character"} give shorthened
versions of the GESIS labels for easier handling in R.
This is a wrapper function around the \code{\link{rescale_categories}} function.
}
\examples{
rescale_occupation (c("Owner of a shop, craftsmen, etc.",
                      "Retired, unable to work",
                      "Employed position, at desk",
                      "Employed professional (employed doctor, etc.)",
                      "DK"),
                    return_class = "character")

##Nicer printing factor names without underscore:

rescale_occupation (c("Owner of a shop, craftsmen, etc.",
                      "Retired, unable to work","DK"),
                    underscore = FALSE)

##Beware that the numeric representation is grouped but not ordered.

rescale_occupation (c("Owner of a shop, craftsmen, etc.",
                      "Retired, unable to work",
                      "Employed position, at desk",
                      "Employed professional (employed doctor, etc.)",
                      "DK"),
                    return_class = "numeric")

}
