% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_time_frequency_use.R
\name{rescale_time_frequency_use}
\alias{rescale_time_frequency_use}
\title{Rescale time frequency of use}
\usage{
rescale_time_frequency_use(column, from = c("Everyday/Almost everyday",
  "Two or three times a week", "About once a week",
  "Two or three times a month", "Less often", "Never"), to = c(250, 120, 50,
  25, 12, 0), exact_from = TRUE, na_labels = c("NA", "DK"),
  return_class = "numeric")
}
\arguments{
\item{column}{A column from a survey data frame where frequency should be
converted to a numeric value.}

\item{from}{Defaults to \code{c("Everyday/Almost everyday",
"Two or three times a week","About once a week",
"Two or three times a month", "Less often", "Never")}.}

\item{to}{Defaults to \code{c(250,120, 50,25,12,0)}.}

\item{exact_from}{Deafults to \code{TRUE}. If \code{FALSE} you can use the
partial matching, but beware that in this case, \code{"twenty"} will be replaced by}

\item{na_labels}{Defaults to \code{c("NA","DK")}.}

\item{return_class}{Default is \code{"numeric"}, alternatives \code{"character"} or
\code{"factor"}.}
}
\description{
Create a uniform variable from use frequency categories such as code{
c("Everyday/Almost everyday", "Two or three times a week",
"About once a week", "Two or three times a month",
"Less often", "Never")}. This is a wrapper function around the
\code{\link{rescale_categories}} function.
}
\examples{
rescale_time_frequency_use (column = c(
 "Everyday/Almost everyday", "Less often", "never", "NEVER", "NA", "DK"),
 na_labels = "default")
}
