% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_value_labels.R
\name{unique_value_labels}
\alias{unique_value_labels}
\title{Summarize unique value labels of a variable.}
\usage{
unique_value_labels(x, numeric_values = FALSE)
}
\arguments{
\item{x}{A variable to summarize.}

\item{numeric_values}{Defaults to \code{FALSE} in which case the integer
and numeric variables are no analyized. If \code{(TRUE)} it will create
a similar character summary of the numeric variables.}
}
\description{
Summarize unique value labels of a variable.
}
\examples{
unique_value_labels (x = c("apple", "strawberry", NA, "apple"))

#By default will not summarize numeric variables
unique_value_labels (x = c(1,2, NA, 3))

#But it is possible to create a similar summary of numeric variables.
unique_value_labels (x = c(1,2, NA, 3),
                     numeric_values = TRUE)
}
