% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values_factor_3_get.R
\name{values_factor_3_get}
\alias{values_factor_3_get}
\title{Get three level (positive, neutral, negative) categories}
\usage{
values_factor_3_get()
}
\description{
Get three level (positive, neutral, negative) categories
}
\examples{
values_factor_3_get()

}
