% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values_factor_4_get.R
\name{values_factor_4_get}
\alias{values_factor_4_get}
\title{Get four level non-negative categories}
\usage{
values_factor_4_get()
}
\description{
Get four level non-negative categories
}
\examples{
values_factor_4_get()

}
