% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values_pos_neg_4_get.R
\name{values_pos_neg_4_get}
\alias{values_pos_neg_4_get}
\title{Get four level (2 positive, 2 negative) categories}
\usage{
values_pos_neg_4_get()
}
\description{
Get four level (2 positive, 2 negative) categories
}
\examples{
values_pos_neg_get()

}
