% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values_pos_neg_get.R
\name{values_pos_neg_get}
\alias{values_pos_neg_get}
\title{Get three level (positive, neutral, negative) categories}
\usage{
values_pos_neg_get()
}
\description{
Get three level (positive, neutral, negative) categories
}
\examples{
values_pos_neg_get()

}
