% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values_yes_no_4_get.R
\name{values_yes_no_4_get}
\alias{values_yes_no_4_get}
\title{Get four-level yes values}
\usage{
values_yes_no_4_get()
}
\description{
Get the character strings that will be recoded as "yes" in
four-level variables.
}
\examples{
values_yes_no_4_get()

}
