% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_name_suggest.R
\name{var_name_suggest}
\alias{var_name_suggest}
\title{Suggest a new variable name}
\usage{
var_name_suggest(x)
}
\arguments{
\item{x}{The original GESIS variable label (long)}
}
\description{
Suggest a new variable name
}
\examples{
\dontrun{
var_name_suggest ( gesis_name )
}
}
