% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary_items_get.R
\name{vocabulary_items_get}
\alias{vocabulary_items_get}
\title{Get the recoding vocabulary}
\usage{
vocabulary_items_get(context_var = NULL)
}
\arguments{
\item{context_var}{Filtering for the vocabulary of a class only, such as the
vocabulary of only 3-level \code{factor_3} variables. Defaults to \code{NULL}.}
}
\description{
Get the vocabulary used for re-coding or reordering
categorical variables
}
\examples{
vocabulary_items_get()
}
