% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-vocabulary_nuts2.R
\docType{data}
\name{vocabulary_nuts2}
\alias{vocabulary_nuts2}
\title{A vocabulary file for NUTS2 region coding}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 239 rows and 4 columns.}
\source{
\url{http://ec.europa.eu/eurostat/documents/345175/629341/NUTS+2010+-+NUTS+2013.xls}
}
\usage{
vocabulary_nuts2
}
\description{
Currently only NUT2 2010 and 2013 codes are included.
\describe{
  \item{country_code}{Country code}
  \item{region_nuts_codes}{Region name as coded by GESIS}
  \item{code2010}{NUTS2 2010 coding for years 2012, 2013, 2014.}
  \item{code2013}{NUTS2 2013 coding for years 2015, 2016, 2017.}
}
}
\keyword{datasets}
