% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yes_4_get.R
\name{yes_4_get}
\alias{yes_4_get}
\title{Get four-level yes-no values}
\usage{
yes_4_get()
}
\description{
Get the character strings that will be recoded as "yes" in
four-level variables.
}
\examples{
yes_4_get()

}
