% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yes_no_4_get.R
\name{yes_no_4_get}
\alias{yes_no_4_get}
\title{Get four-level yes values}
\usage{
yes_no_4_get()
}
\description{
Get the character strings that will be recoded as "yes" in
four-level variables.
}
\examples{
yes_no_4_get

}
