% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zacat_file_lookup.R
\name{zacat_file_lookup}
\alias{zacat_file_lookup}
\title{Look up a ZACAT file in a directory}
\usage{
zacat_file_lookup(zacat_id = NULL, data_dir = NULL, see_log = TRUE,
  create_log = TRUE, my_treshold = futile.logger::INFO)
}
\arguments{
\item{zacat_id}{\code{NULL}}

\item{data_dir}{\code{NULL}}

\item{see_log}{\code{TRUE} which will print messages to the screen.}

\item{create_log}{It will create log files in the sr_logs director.}

\item{my_treshold}{Can be \code{futile.logger::WARN},
\code{futile.logger::INFO}, \code{futile.logger::ERROR}.Defaults to
\code{futile.logger::INFO}.}
}
\description{
Tries to find a data file based on ZACAT ID  on a data directory path,
or, if omitted, on the current \code{tempdir()}.
}
\examples{
\dontrun{
##use your own file path:
zacat_file_lookup ( zacat_id = "ZA4744",
                data_dir = data_dir,
                see_log = see_log,
                create_log = creat_log,
                my_treshold = my_treshold )
}
}
