% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_factor_frequency.R
\name{as_factor_frequency}
\alias{as_factor_frequency}
\title{Constructor for quasi-numeric frequency factors}
\usage{
as_factor_frequency(x)
}
\arguments{
\item{x}{An input vector to be converted to a
four non-negative level variable}
}
\description{
Constructor for quasi-numeric frequency factors
}
\examples{
as_factor_4( c("Not at all", "Low",
               "Medium", "Strong", "DK"))

}
