% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesis_file_read.R
\name{gesis_file_read}
\alias{gesis_file_read}
\title{Import a GESIS SPSS file}
\usage{
gesis_file_read(path, new_names = NULL, create_metadata_file = TRUE)
}
\arguments{
\item{path}{The path to the GESIS SPSS file.}

\item{new_names}{Defaults to \code{NULL}.}

\item{create_metadata_file}{If \code{TRUE} it will create
an .rds file with the metadata of the GESIS file.}
}
\description{
Import a GESIS SPSS file
}
\examples{
\dontrun{
ZA5929 <- gesis_file_read("data-raw/ZA5929_v3-0-0.sav")
}
}
