% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.factor_pos_neg.R, R/is.factor_yes_no_4.R
\name{is.factor_pos_neg}
\alias{is.factor_pos_neg}
\alias{is.factor_pos_neg}
\title{Is the object of class factor_pos_neg?}
\usage{
is.factor_pos_neg(x)

is.factor_pos_neg(x)
}
\arguments{
\item{x}{An object to be checked if it is a 3-level (positive,
neutral, negative) categorical variable.}

\item{x}{An object to be checked if it is a 4-level (2 positive,
2 negative) categorical variable.}
}
\description{
Is the object of class factor_pos_neg?

Is the object of class factor_yes_no_4?
}
\examples{
my_categories <- as_factor_pos_neg( c("Better", "DK", "Worse",
                               "Same", "The Same","Inap. not "))

is.factor_pos_neg (my_categories)

my_categories <- as_factor_pos_neg( c("Better", "DK", "Worse",
                               "Same", "The Same","Inap. not "))

is.factor_pos_neg (my_categories)

}
