% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary_items_get.R
\name{vocabulary_items_get}
\alias{vocabulary_items_get}
\title{Get the recoding vocabulary}
\usage{
vocabulary_items_get(type = NULL)
}
\description{
Get the vocabulary used for re-coding or reordering
categorical variables
}
\examples{
vocabulary_items_get()
}
